Manually invoke qmail-inject and qmail-queue separately.
By using qmail-inject only as a message filter, we can fully control the
envelope (sender and recipient addresses) sent to qmail queue instead of
demanding that qmail-inject parse the header.  Further, we now capture
and output the qmail PID (qp) number so that it will show up in logs for
subsequent tracing.

This change significantly changes the internals of how messages are
sent, but should make no actual differences in behavior or in the
resulting message.